[API] Move remaining IProject methods to IContainer
Moves the remaining IProject methods getFile(...) and getFolder(...)
into IContainer. This is done as a preparation for the migration to a
reference point based resource system where the project metaphor is
replaced by a base folder everything is described relative to.

With this change, IProject no longer contains any methods and only
functions as a marker interface with the current logic. The scope of the
interface might be extended later on after it was repurposed for the new
reference point model.

Simply moves the implementation to the abstract IContainer
implementation class for the Eclipse and Server implementation.

Copies the existing logic in IntelliJProjectImpl into
IntellijFolderImpl. This code duplication is currently necessary as
there is no real possibility for a common superclass for IContainer
implementations with the current implementations. This logic should be
unified once the resource model has been overhauled.
